requirements.txt
fastapi>=0.115.2
uvicorn>=0.31.1
pymongo==4.5.0
python-dotenv>=1.1.0
bcrypt==4.1.1
pydantic>=2.11.0
pydantic-settings>=2.5.2
python-multipart>=0.0.18
email-validator==2.3.0
anyio>=4.7.0
PyJWT==2.8.0
